feat(app)!: open web gui automatically in the browser
BREAKING CHANGES: makes web an optional Cargo feature
