+++ bumblebee/Cargo.toml
@@ -15,7 +15,9 @@ name = "bumblebee"
 path = "src/main.rs"
 
 [features]
+default = []
 console = ["dep:console-subscriber"]
+web = ["dep:axum", "dep:webbrowser", "dep:serde"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -29,8 +31,9 @@ soundpad-remote-client = { path = "../client", version = "0.2.0" }
 clap = { version = "4.0.29", features = ["derive"] }
 tracing-subscriber = "0.3.16"
 console-subscriber = { version = "0.1.8", optional = true }
-axum = "0.6.1"
-serde = { version = "1.0.149", features = ["derive"] }
+serde = { version = "1.0.149", features = ["derive"], optional = true }
+axum = { version = "0.6.1", optional = true }
+webbrowser = { version = "0.8.2", optional = true }
 
 [package.metadata.workspaces]
 independent = true
